Update 01_qc_and_trimming.nf
Replaced Nanofilt with Chopper. 

Chopper is a tool that combines the utility of NanoFilt and NanoLyse, for filtering sequencing reads based on quality, length, and contaminating sequences, delivers a 7-fold speed up compared to the Python implementation, making use of the Rust-Bio library (Köster 2016) and Rust bindings to minimap2 (Li 2018).